requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
